Fixed a couple docstrings
